# openapi.yaml
openapi: 3.1.0
info:
  title: Leave Management System API
  version: 1.0.0
  description: |
    A comprehensive API for managing employee leave requests, work-from-home requests,
    holiday planning, and document generation. This API supports role-based access control
    with different permissions for Employees, Managers, HR, Executives, and Admins.

    ## Authentication
    All endpoints (except `/api/health`) require authentication via NextAuth JWT sessions.
    Authentication is handled via Azure AD OAuth 2.0.

    ## Authorization
    API endpoints are protected by role-based access control (RBAC):
    - **EMPLOYEE**: Basic access to own requests and team calendar
    - **MANAGER**: Can approve team member requests
    - **DEPARTMENT_DIRECTOR**: Department-level oversight
    - **HR**: User management, document verification
    - **EXECUTIVE**: Company-wide analytics and approvals
    - **ADMIN**: Full system configuration access

  contact:
    name: API Support
    email: support@company.com
  license:
    name: Proprietary
    url: https://company.com/license
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://leave.company.com
    description: Production server
...
security:
  - bearerAuth: []
paths:
...
  /api/wfh-requests/{id}/self-cancel:
    post:
      tags:
        - WFH Requests
      summary: Self-cancel WFH request
      operationId: selfCancelWfhRequest
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: WFH request cancelled
        "401":
          $ref: "#/components/responses/Unauthorized"

  # Manager Endpoints
...
  /api/setup/auth:
    post:
      tags:
        - Setup
      summary: Authenticate for setup
      description: Authenticates with SETUP_PASSWORD to access setup endpoints
      operationId: setupAuth
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - password
              properties:
                password:
                  type: string
      responses:
        "200":
          description: Authentication successful
        "401":
          description: Invalid password
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: NextAuth JWT session token
    cronSecret:
      type: http
      scheme: bearer
      description: CRON_SECRET environment variable for scheduled tasks
...
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: Unauthorized
...
  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
...